Fix hashmap implementation
Originally, the hashmap was built with VLA's that weren't boundary
aligned. This was causing issues, since we weren't able to index into
the buckets properly without over-running data.

This is fixed by removing VLA's on the heap, and instead emulating them
by over-allocating and indexing with a custom method. This new method
takes into account item alignment, as well as bucket alignment, to
ensure that we are always aligned on a boundary so we can modify data as
needed. This refactor fixes the bug, but it does make accessing all
allocated items difficult.